Initial commit. Not yet tested.
I've edited Rahul Kumar's XenServer VM Backup script to detect a failed mount, log the failure, and exit without backing up. I added this after a mount failed, and it ended up backing up my VMs to the hypervisor's local disk, filling the disk and slowing the hypervisor to a crawl. 

This script runs occasionally to look for these failure logs, and if found, emails the names and dates to me.