Innitial, toothless upload of xenbackup_cleanup.sh
Like I said in my opening comments, I've had automatic VM backups running for a while now, and I'm tired of manually deleting the old ones. I tried to write everything in an adaptable and generic manner, so that anyone can copy me, change 3 variables at the top, and be off to the races. We all get an immense amount of value from free code from people around the world, so these lame scripts are my attempt at repayment; take them and save an hour writing this yourself.

I plan to test this script today and Monday, then give it teeth and set it loose on my backups on Tuesday. Once I'm happy with it, I'll upload it here as a 1.0 sable release.